+++ eth_defi/uniswap_v3/utils.py
@@ -74,7 +74,7 @@ def decode_path(full_path_encoded: bytes) -> list:
     """
 
     assert type(full_path_encoded == bytes), "encoded path must be provided as bytes"
-    
+
     path_pos = 0
     full_path_decoded = []
     # read alternating 20 and 3 byte chunks from the encoded path,