+++ tests/test_decode_tx.py
@@ -9,24 +9,28 @@ To run tests in this module:
 
 """
 import os
+import logging
+import shutil
 
-import flaky
 import pytest
+
 from eth_account import Account
 from eth_account.signers.local import LocalAccount
 from eth_typing import HexAddress, HexStr
 from web3 import HTTPProvider, Web3
 
-from eth_defi.ganache import fork_network
+from eth_defi.anvil import fork_network_anvil
+from eth_defi.chain import install_chain_middleware
+from eth_defi.gas import node_default_gas_price_strategy
 from eth_defi.hotwallet import HotWallet
 from eth_defi.token import fetch_erc20_details
-
-# https://docs.pytest.org/en/latest/how-to/skipping.html#skip-all-test-functions-of-a-class-or-module
 from eth_defi.tx import decode_signed_transaction
 
+
+# https://docs.pytest.org/en/latest/how-to/skipping.html#skip-all-test-functions-of-a-class-or-module
 pytestmark = pytest.mark.skipif(
-    os.environ.get("BNB_CHAIN_JSON_RPC") is None,
-    reason="Set BNB_CHAIN_JSON_RPC environment variable to Binance Smart Chain node to run this test",
+    (os.environ.get("BNB_CHAIN_JSON_RPC") is None) or (shutil.which("anvil") is None),
+    reason="Set BNB_CHAIN_JSON_RPC env install anvil command to run these tests",
 )
 
 
@@ -48,24 +52,29 @@ def user_1() -> LocalAccount:
     return Account.create()
 
 
-@pytest.fixture(scope="module")
-def ganache_bnb_chain_fork(large_busd_holder) -> str:
+@pytest.fixture()
+def anvil_bnb_chain_fork(request, large_busd_holder) -> str:
     """Create a testable fork of live BNB chain.
 
     :return: JSON-RPC URL for Web3
     """
     mainnet_rpc = os.environ["BNB_CHAIN_JSON_RPC"]
-    launch = fork_network(mainnet_rpc, unlocked_addresses=[large_busd_holder])
-    yield launch.json_rpc_url
-    # Wind down Ganache process after the test is complete
-    launch.close()
+    launch = fork_network_anvil(mainnet_rpc, unlocked_addresses=[large_busd_holder])
+    try:
+        yield launch.json_rpc_url
+    finally:
+        # Wind down Anvil process after the test is complete
+        launch.close(log_level=logging.ERROR)
 
 
 @pytest.fixture
-def web3(ganache_bnb_chain_fork: str):
+def web3(anvil_bnb_chain_fork: str):
     """Set up a local unit testing blockchain."""
     # https://web3py.readthedocs.io/en/stable/examples.html#contract-unit-tests-in-python
-    return Web3(HTTPProvider(ganache_bnb_chain_fork))
+    web3 = Web3(HTTPProvider(anvil_bnb_chain_fork))
+    install_chain_middleware(web3)
+    web3.eth.set_gas_price_strategy(node_default_gas_price_strategy)
+    return web3
 
 
 @pytest.fixture
@@ -77,7 +86,6 @@ def hot_wallet(user_1: LocalAccount, web3: Web3) -> HotWallet:
     return wallet
 
 
-@flaky.flaky(max_runs=5)
 def test_bnb_chain_decode_tx(web3: Web3, large_busd_holder: HexAddress, hot_wallet: HotWallet):
     """Decoding transactions targeting BNB chain."""
 
@@ -87,10 +95,9 @@ def test_bnb_chain_decode_tx(web3: Web3, large_busd_holder: HexAddress, hot_wall
     busd = busd_details.contract
 
     # Create a spoofed transfer() (never executed)
-    raw_tx = busd.functions.transfer("0x0000000000000000000000000000000000000000", 500 * 10**18).buildTransaction({"gas": 100_000})
+    raw_tx = busd.functions.transfer("0x0000000000000000000000000000000000000000", 500 * 10**18).build_transaction({"gas": 100_000})
     signed_tx = hot_wallet.sign_transaction_with_new_nonce(raw_tx)
     signed_tx_bytes = signed_tx.rawTransaction
     d = decode_signed_transaction(signed_tx_bytes)
-    assert d["chainId"] == 1337
     assert d["nonce"] == 0
     assert d["data"].hex().startswith("0xa9059cbb0")  # transfer() function selector