+++ CHANGELOG.md
@@ -1,10 +1,15 @@
 # Current
 
-# 0.13.12
-
+- Replace `ganache` with `anvil` for mainnet fork solution. Anvil is much more stable
+  than Ganache what comes to JSON-RPC. Anvil is much faster. You can now call
+  [fork_network_anvil]() that is a drop-in replacement for old
+  Ganache based `fork_network`.
+- Move internal test suite to use Anvil instead of Ganache. This allows us to remove
+  `flaky` decorators on tests.
 - Add `analysis.py` for Uniswap V3 and relevant tests
 - Add `mock_partial_deployment` function for V3
 - Abstract `TradeResult`, `TradeSuccess`, and `TradeFailure` out of Uniswap V2 and into eth_defi.trade, since also used in Uniswap V3
+
 
 # 0.13.11